增加直播检测demo和修改文本检测文件名
